update source file
This version is for testing only. This testing is implemented to see whether dynamic reconfigure will overwrite whatever parameter that is in the server. Depending on when you set the callback function, it will run the callback function and reset the params as you told it to do. Therefore, if you do not wish to reset the params you loading into your node, please call perform the set callback function first and then only the param read function.